# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Music App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Track {
  id: string;
  title: string;
  artist: string;
  album: string;
  /** Duration in milliseconds */
  duration: number;
  /** @nullable */
  previewUrl: string | null;
  artworkUrl: string;
  /** @nullable */
  genre?: string | null;
  /** @nullable */
  releaseYear?: number | null;
}

export interface Playlist {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  trackCount: number;
  /** @nullable */
  coverArt?: string | null;
  createdAt: string;
}

export interface PlaylistWithTracks {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  trackCount: number;
  /** @nullable */
  coverArt?: string | null;
  createdAt: string;
  tracks: Track[];
}

export interface PlaylistInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface PlaylistUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
}

export interface PlaylistTrackInput {
  trackId: string;
  title: string;
  artist: string;
  album: string;
  duration: number;
  /** @nullable */
  previewUrl?: string | null;
  artworkUrl: string;
  /** @nullable */
  genre?: string | null;
  /** @nullable */
  releaseYear?: number | null;
}

export interface ErrorResponse {
  error: string;
}

export type SearchMusicParams = {
/**
 * Search query (artist, song title, etc.)
 */
q: string;
/**
 * Max results (default 25)
 */
limit?: number;
};

export type GetTopTracksParams = {
/**
 * Genre filter
 */
genre?: string;
};