# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Music App API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: music
    description: Music search and metadata
  - name: playlists
    description: Playlist management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /music/search:
    get:
      operationId: searchMusic
      tags: [music]
      summary: Search for music tracks
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query (artist, song title, etc.)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
          description: Max results (default 25)
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"

  /music/top:
    get:
      operationId: getTopTracks
      tags: [music]
      summary: Get top/featured tracks
      parameters:
        - name: genre
          in: query
          required: false
          schema:
            type: string
          description: Genre filter
      responses:
        "200":
          description: Top tracks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"

  /music/track/{trackId}:
    get:
      operationId: getTrack
      tags: [music]
      summary: Get a specific track by ID
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Track details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "404":
          description: Track not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /playlists:
    get:
      operationId: getPlaylists
      tags: [playlists]
      summary: List all playlists
      responses:
        "200":
          description: List of playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playlist"
    post:
      operationId: createPlaylist
      tags: [playlists]
      summary: Create a new playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistInput"
      responses:
        "201":
          description: Created playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"

  /playlists/{playlistId}:
    get:
      operationId: getPlaylist
      tags: [playlists]
      summary: Get a playlist with its tracks
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Playlist with tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistWithTracks"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePlaylist
      tags: [playlists]
      summary: Rename or update a playlist
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistUpdate"
      responses:
        "200":
          description: Updated playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
    delete:
      operationId: deletePlaylist
      tags: [playlists]
      summary: Delete a playlist
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /playlists/{playlistId}/tracks:
    post:
      operationId: addTrackToPlaylist
      tags: [playlists]
      summary: Add a track to a playlist
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistTrackInput"
      responses:
        "201":
          description: Track added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistWithTracks"

  /playlists/{playlistId}/tracks/{trackId}:
    delete:
      operationId: removeTrackFromPlaylist
      tags: [playlists]
      summary: Remove a track from a playlist
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistWithTracks"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Track:
      type: object
      required:
        - id
        - title
        - artist
        - album
        - duration
        - previewUrl
        - artworkUrl
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        duration:
          type: integer
          description: Duration in milliseconds
        previewUrl:
          type: ["string", "null"]
        artworkUrl:
          type: string
        genre:
          type: ["string", "null"]
        releaseYear:
          type: ["integer", "null"]

    Playlist:
      type: object
      required:
        - id
        - name
        - trackCount
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        trackCount:
          type: integer
        coverArt:
          type: ["string", "null"]
        createdAt:
          type: string

    PlaylistWithTracks:
      type: object
      required:
        - id
        - name
        - trackCount
        - createdAt
        - tracks
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        trackCount:
          type: integer
        coverArt:
          type: ["string", "null"]
        createdAt:
          type: string
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"

    PlaylistInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string

    PlaylistUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string

    PlaylistTrackInput:
      type: object
      required:
        - trackId
        - title
        - artist
        - album
        - duration
        - artworkUrl
      properties:
        trackId:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        duration:
          type: integer
        previewUrl:
          type: ["string", "null"]
        artworkUrl:
          type: string
        genre:
          type: ["string", "null"]
        releaseYear:
          type: ["integer", "null"]

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string