# lib/api-zod/src/generated/types/playlist.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Music App API
 * OpenAPI spec version: 0.1.0
 */

export interface Playlist {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  trackCount: number;
  /** @nullable */
  coverArt?: string | null;
  createdAt: string;
}