# lib/api-zod/src/generated/types/playlistTrackInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Music App API
 * OpenAPI spec version: 0.1.0
 */

export interface PlaylistTrackInput {
  trackId: string;
  title: string;
  artist: string;
  album: string;
  duration: number;
  /** @nullable */
  previewUrl?: string | null;
  artworkUrl: string;
  /** @nullable */
  genre?: string | null;
  /** @nullable */
  releaseYear?: number | null;
}