# lib/api-zod/src/generated/types/track.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Music App API
 * OpenAPI spec version: 0.1.0
 */

export interface Track {
  id: string;
  title: string;
  artist: string;
  album: string;
  /** Duration in milliseconds */
  duration: number;
  /** @nullable */
  previewUrl: string | null;
  artworkUrl: string;
  /** @nullable */
  genre?: string | null;
  /** @nullable */
  releaseYear?: number | null;
}